+++ capi/example/simulator_server.c
@@ -4,7 +4,7 @@
  * Created Date: 10/10/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 11/10/2022
+ * Last Modified: 21/12/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -13,7 +13,7 @@
 
 #include "simulator.h"
 
-int main() {
+int main(void) {
   const char* settings = "settings.json";
 
   AUTDExtraSimulator(settings, true, 0);