+++ cs/src/AUTD3Sharp.cs
@@ -4,7 +4,7 @@
  * Created Date: 23/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 16/12/2022
+ * Last Modified: 24/12/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -13,7 +13,7 @@
 
 #if UNITY_2018_3_OR_NEWER
 #define USE_SINGLE
-#define DIMENSION_M
+#define DIMENSION_M 
 #endif
 
 using Microsoft.Win32.SafeHandles;
@@ -35,6 +35,11 @@ using Quaternion = AUTD3Sharp.Utils.Quaterniond;
 using System.Numerics;
 #endif
 
+#if USE_SINGLE
+using autd3_float_t = System.Single;
+#else
+using autd3_float_t = System.Double;
+#endif
 
 namespace AUTD3Sharp
 {
@@ -98,6 +103,7 @@ namespace AUTD3Sharp
         public const byte DriverV2_4 = 0x84;
         public const byte DriverV2_5 = 0x85;
         public const byte DriverV2_6 = 0x86;
+        public const byte DriverV2_7 = 0x87;
 
         #endregion
 
@@ -113,40 +119,6 @@ namespace AUTD3Sharp
             var onOutput = Marshal.GetFunctionPointerForDelegate(output);
             var onFlush = Marshal.GetFunctionPointerForDelegate(flush);
             Base.AUTDSetDefaultLogger(onOutput, onFlush);
-        }
-    }
-
-    public static class TypeHelper
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static (double, double, double) Convert(Vector3 vector)
-        {
-#if USE_SINGLE
-            return ((double)vector.x, (double)vector.y, (double)vector.z);
-#else
-            return (vector.x, vector.y, vector.z);
-#endif
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static Vector3 Convert(double x, double y, double z)
-        {
-#if USE_SINGLE
-            var vector = new Vector3((float)x, (float)y, (float)z);
-#else
-            var vector = new Vector3(x, y, z);
-#endif
-            return vector;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static (double, double, double, double) Convert(Quaternion quaternion)
-        {
-#if USE_SINGLE
-            return ((double)quaternion.w, (double)quaternion.x, (double)quaternion.y, (double)quaternion.z);
-#else
-            return (quaternion.w, quaternion.x, quaternion.y, quaternion.z);
-#endif
         }
     }
 
@@ -168,7 +140,7 @@ namespace AUTD3Sharp
             get
             {
                 Base.AUTDTransPosition(_cnt, _trId, out var x, out var y, out var z);
-                return TypeHelper.Convert(x, y, z);
+                return new Vector3(x, y, z);
             }
         }
 
@@ -177,7 +149,7 @@ namespace AUTD3Sharp
             get
             {
                 Base.AUTDTransXDirection(_cnt, _trId, out var x, out var y, out var z);
-                return TypeHelper.Convert(x, y, z);
+                return new Vector3(x, y, z);
             }
         }
 
@@ -186,7 +158,7 @@ namespace AUTD3Sharp
             get
             {
                 Base.AUTDTransYDirection(_cnt, _trId, out var x, out var y, out var z);
-                return TypeHelper.Convert(x, y, z);
+                return new Vector3(x, y, z);
             }
         }
 
@@ -195,13 +167,13 @@ namespace AUTD3Sharp
             get
             {
                 Base.AUTDTransZDirection(_cnt, _trId, out var x, out var y, out var z);
-                return TypeHelper.Convert(x, y, z);
+                return new Vector3(x, y, z);
             }
         }
 
-        public double Wavelength => Base.AUTDGetWavelength(_cnt, _trId);
+        public autd3_float_t Wavelength => Base.AUTDGetWavelength(_cnt, _trId);
 
-        public double Frequency
+        public autd3_float_t Frequency
         {
             get => Base.AUTDGetTransFrequency(_cnt, _trId);
             set => Base.AUTDSetTransFrequency(_cnt, _trId, value);
@@ -229,19 +201,9 @@ namespace AUTD3Sharp
             CntPtr = cntPtr;
         }
 
-        public void AddDevice(Vector3 position, Vector3 rotation)
-        {
-            var (x, y, z) = TypeHelper.Convert(position);
-            var (rx, ry, rz) = TypeHelper.Convert(rotation);
-            Base.AUTDAddDevice(CntPtr, x, y, z, rx, ry, rz);
-        }
+        public void AddDevice(Vector3 position, Vector3 rotation) => Base.AUTDAddDevice(CntPtr, position.x, position.y, position.z, rotation.x, rotation.y, rotation.z);
 
-        public void AddDevice(Vector3 position, Quaternion quaternion)
-        {
-            var (x, y, z) = TypeHelper.Convert(position);
-            var (qw, qx, qy, qz) = TypeHelper.Convert(quaternion);
-            Base.AUTDAddDeviceQuaternion(CntPtr, x, y, z, qw, qx, qy, qz);
-        }
+        public void AddDevice(Vector3 position, Quaternion quaternion) => Base.AUTDAddDeviceQuaternion(CntPtr, position.x, position.y, position.z, quaternion.w, quaternion.x, quaternion.y, quaternion.z);
 
         public int NumTransducers => Base.AUTDNumTransducers(CntPtr);
 
@@ -252,7 +214,7 @@ namespace AUTD3Sharp
             get
             {
                 Base.AUTDGeometryCenter(CntPtr, out var x, out var y, out var z);
-                return TypeHelper.Convert(x, y, z);
+                return new Vector3(x, y, z);
             }
         }
 
@@ -373,13 +335,13 @@ namespace AUTD3Sharp
         public Geometry Geometry { get; }
 
 
-        public double SoundSpeed
+        public autd3_float_t SoundSpeed
         {
             get => Base.AUTDGetSoundSpeed(AUTDControllerHandle.CntPtr);
             set => Base.AUTDSetSoundSpeed(AUTDControllerHandle.CntPtr, value);
         }
 
-        public double Attenuation
+        public autd3_float_t Attenuation
         {
             get => Base.AUTDGetAttenuation(AUTDControllerHandle.CntPtr);
             set => Base.AUTDSetAttenuation(AUTDControllerHandle.CntPtr, value);
@@ -434,7 +396,7 @@ namespace AUTD3Sharp
         #endregion
 
 
-        public void SetSoundSpeedFromTemp(double temp, double k = 1.4, double r = 8.31446261815324, double m = 28.9647e-3)
+        public void SetSoundSpeedFromTemp(autd3_float_t temp, autd3_float_t k = (autd3_float_t)1.4, autd3_float_t r = (autd3_float_t)8.31446261815324, autd3_float_t m = (autd3_float_t)28.9647e-3)
         {
             Base.AUTDSetSoundSpeedFromTemp(AUTDControllerHandle.CntPtr, temp, k, r, m);
         }
@@ -602,7 +564,7 @@ namespace AUTD3Sharp
 
     public sealed class Amplitudes : Body
     {
-        public Amplitudes(double amp = 1.0)
+        public Amplitudes(autd3_float_t amp = (autd3_float_t)1.0)
         {
             Base.AUTDCreateAmplitudes(out handle, amp);
         }
@@ -633,11 +595,7 @@ namespace AUTD3Sharp
 
         public sealed class Focus : Gain
         {
-            public Focus(Vector3 point, double amp = 1.0)
-            {
-                var (x, y, z) = TypeHelper.Convert(point);
-                Base.AUTDGainFocus(out handle, x, y, z, amp);
-            }
+            public Focus(Vector3 point, autd3_float_t amp = (autd3_float_t)1.0) => Base.AUTDGainFocus(out handle, point.x, point.y, point.z, amp);
         }
 
         public sealed class Grouped : Gain
@@ -655,26 +613,17 @@ namespace AUTD3Sharp
 
         public sealed class BesselBeam : Gain
         {
-            public BesselBeam(Vector3 point, Vector3 dir, double thetaZ, double amp = 1.0)
-            {
-                var (x, y, z) = TypeHelper.Convert(point);
-                var (dx, dy, dz) = TypeHelper.Convert(dir);
-                Base.AUTDGainBesselBeam(out handle, x, y, z, dx, dy, dz, thetaZ, amp);
-            }
+            public BesselBeam(Vector3 point, Vector3 dir, autd3_float_t thetaZ, autd3_float_t amp = (autd3_float_t)1.0) => Base.AUTDGainBesselBeam(out handle, point.x, point.y, point.z, dir.x, dir.y, dir.z, thetaZ, amp);
         }
 
         public sealed class PlaneWave : Gain
         {
-            public PlaneWave(Vector3 dir, double amp = 1.0)
-            {
-                var (dx, dy, dz) = TypeHelper.Convert(dir);
-                Base.AUTDGainPlaneWave(out handle, dx, dy, dz, amp);
-            }
+            public PlaneWave(Vector3 dir, autd3_float_t amp = (autd3_float_t)1.0) => Base.AUTDGainPlaneWave(out handle, dir.x, dir.y, dir.z, amp);
         }
 
         public sealed class Custom : Gain
         {
-            public Custom(double[] amp, double[] phase)
+            public Custom(autd3_float_t[] amp, autd3_float_t[] phase)
             {
                 if (amp.Length != phase.Length) throw new ArgumentException();
                 var length = amp.Length;
@@ -707,7 +656,7 @@ namespace AUTD3Sharp
                 return true;
             }
 
-            public double SamplingFrequency => Base.AUTDModulationSamplingFrequency(handle);
+            public autd3_float_t SamplingFrequency => Base.AUTDModulationSamplingFrequency(handle);
             public uint SamplingFrequencyDivision
             {
                 get => Base.AUTDModulationSamplingFrequencyDivision(handle);
@@ -717,7 +666,7 @@ namespace AUTD3Sharp
 
         public sealed class Static : Modulation
         {
-            public Static(double amp = 1.0)
+            public Static(autd3_float_t amp = (autd3_float_t)1.0)
             {
                 Base.AUTDModulationStatic(out handle, amp);
 
@@ -726,7 +675,7 @@ namespace AUTD3Sharp
 
         public sealed class Sine : Modulation
         {
-            public Sine(int freq, double amp = 1.0, double offset = 0.5)
+            public Sine(int freq, autd3_float_t amp = (autd3_float_t)1.0, autd3_float_t offset = (autd3_float_t)0.5)
             {
                 Base.AUTDModulationSine(out handle, freq, amp, offset);
             }
@@ -734,7 +683,7 @@ namespace AUTD3Sharp
 
         public sealed class SineSquared : Modulation
         {
-            public SineSquared(int freq, double amp = 1.0, double offset = 0.5)
+            public SineSquared(int freq, autd3_float_t amp = (autd3_float_t)1.0, autd3_float_t offset = (autd3_float_t)0.5)
             {
                 Base.AUTDModulationSineSquared(out handle, freq, amp, offset);
             }
@@ -742,7 +691,7 @@ namespace AUTD3Sharp
 
         public sealed class SineLegacy : Modulation
         {
-            public SineLegacy(double freq, double amp = 1.0, double offset = 0.5)
+            public SineLegacy(autd3_float_t freq, autd3_float_t amp = (autd3_float_t)1.0, autd3_float_t offset = (autd3_float_t)0.5)
             {
                 Base.AUTDModulationSineLegacy(out handle, freq, amp, offset);
             }
@@ -751,7 +700,7 @@ namespace AUTD3Sharp
 
         public sealed class Square : Modulation
         {
-            public Square(int freq, double low = 0.0, double high = 1.0, double duty = 0.5)
+            public Square(int freq, autd3_float_t low = (autd3_float_t)0.0, autd3_float_t high = (autd3_float_t)1.0, autd3_float_t duty = (autd3_float_t)0.5)
             {
                 Base.AUTDModulationSquare(out handle, freq, low, high, duty);
             }
@@ -776,7 +725,7 @@ namespace AUTD3Sharp
                 return true;
             }
 
-            public double Frequency
+            public autd3_float_t Frequency
             {
                 get => Base.AUTDSTMFrequency(handle);
                 set => Base.AUTDSTMSetFrequency(handle, value);
@@ -794,7 +743,7 @@ namespace AUTD3Sharp
                 set => Base.AUTDSTMSetFinishIdx(handle, value);
             }
 
-            public double SamplingFrequency => Base.AUTDSTMSamplingFrequency(handle);
+            public autd3_float_t SamplingFrequency => Base.AUTDSTMSamplingFrequency(handle);
             public uint SamplingFrequencyDivision
             {
                 get => Base.AUTDSTMSamplingFrequencyDivision(handle);
@@ -804,16 +753,12 @@ namespace AUTD3Sharp
 
         public sealed class FocusSTM : STM
         {
-            public FocusSTM(double soundSpeed)
+            public FocusSTM(autd3_float_t soundSpeed)
             {
                 Base.AUTDFocusSTM(out handle, soundSpeed);
             }
 
-            public void Add(Vector3 point, byte shift = 0)
-            {
-                var (x, y, z) = TypeHelper.Convert(point);
-                Base.AUTDFocusSTMAdd(handle, x, y, z, shift);
-            }
+            public void Add(Vector3 point, byte shift = 0) => Base.AUTDFocusSTMAdd(handle, point.x, point.y, point.z, shift);
         }
 
         public enum Mode : ushort